+++ assets/javascripts/kitten/components/buttons/button/button.stories.js
@@ -1,6 +1,13 @@
 import React from 'react'
 import { text, boolean, number, radios, select } from '@storybook/addon-knobs'
-import { Button } from './index'
+import {
+  Button,
+  buttonModifiers,
+  buttonFitOptions,
+  buttonMobileFitOptions,
+  buttonSizes,
+  buttonVariants,
+} from './index'
 import { Marger } from '../../layout/marger'
 import { Container } from '../../grid/container'
 import { Grid, GridCol } from '../../grid/grid'
@@ -28,49 +35,16 @@ const svgPositionOptions = {
   Right: 'right',
 }
 
-const variantOptions = {
-  Andromeda: 'andromeda',
-  Orion: 'orion',
-}
-
-const modifierOptions = {
-  Hydrogen: 'hydrogen',
-  Helium: 'helium',
-  Lithium: 'lithium',
-  Beryllium: 'beryllium',
-  Carbon: 'carbon',
-  Oxygen: 'oxygen',
-  Checked: 'checked',
-  Copper: 'copper',
-  Boron: 'boron',
-  Neon: 'neon',
-  Iron: 'iron',
-  Social_facebook: 'social_facebook',
-  Social_twitter: 'social_twitter',
-  Social_linkedin: 'social_linkedin',
-  Social_instagram: 'social_instagram',
-  Social_youtube: 'social_youtube',
-  Social_pinterest: 'social_pinterest',
-}
-
-const sizeOptions = {
-  Nano: 'nano',
-  Micro: 'micro',
-  Tiny: 'tiny',
-  Regular: 'regular',
-  Big: 'big',
-  Huge: 'huge',
-  Giant: 'giant',
-}
-
 export const WithText = () => (
   <Button
-    size={select('Size', sizeOptions, 'regular')}
+    size={select('Size', buttonSizes, 'regular')}
     fluid={boolean('Fluid', false)}
-    modifier={select('Modifier', modifierOptions, 'hydrogen')}
-    variant={select('Variant', variantOptions, 'andromeda')}
+    modifier={select('Modifier', buttonModifiers, 'hydrogen')}
+    variant={select('Variant', buttonVariants, 'andromeda')}
     disabled={boolean('Disabled', false)}
     borderRadius={number('Border radius', 0, radiusBorderRange)}
+    fit={select('fit', buttonFitOptions, 'min-width')}
+    mobileFit={select('mobileFit', buttonMobileFitOptions, null)}
   >
     {text('Text', 'MyButton')}
   </Button>
@@ -92,13 +66,15 @@ WithText.decorators = [
 
 export const WithIcon = () => (
   <Button
-    size={select('Size', sizeOptions, 'regular')}
+    size={select('Size', buttonSizes, 'regular')}
     rounded={boolean('Rounded', false)}
-    modifier={select('Modifier', modifierOptions, 'hydrogen')}
-    variant={select('Variant', variantOptions, 'andromeda')}
+    modifier={select('Modifier', buttonModifiers, 'hydrogen')}
+    variant={select('Variant', buttonVariants, 'andromeda')}
     fluid={boolean('Fluid', false)}
     disabled={boolean('Disabled', false)}
     borderRadius={number('Border radius', 0, radiusBorderRange)}
+    fit={select('fit', buttonFitOptions, 'min-width')}
+    mobileFit={select('mobileFit', buttonMobileFitOptions, null)}
     icon
   >
     <HeartIcon
@@ -124,13 +100,15 @@ WithIcon.decorators = [
 
 export const WithColorIcon = () => (
   <Button
-    size={select('Size', sizeOptions, 'regular')}
+    size={select('Size', buttonSizes, 'regular')}
     rounded={boolean('Rounded', false)}
-    modifier={select('Modifier', modifierOptions, 'hydrogen')}
-    variant={select('Variant', variantOptions, 'andromeda')}
+    modifier={select('Modifier', buttonModifiers, 'hydrogen')}
+    variant={select('Variant', buttonVariants, 'andromeda')}
     fluid={boolean('Fluid', false)}
     disabled={boolean('Disabled', false)}
     borderRadius={number('Border radius', 0, radiusBorderRange)}
+    fit={select('fit', buttonFitOptions, 'min-width')}
+    mobileFit={select('mobileFit', buttonMobileFitOptions, null)}
   >
     <PayPalIcon />
   </Button>
@@ -154,12 +132,14 @@ export const WithTextAndIcon = () => {
   const iconPosition = radios('Icon position', svgPositionOptions, 'left')
   return (
     <Button
-      size={select('Size', sizeOptions, 'regular')}
+      size={select('Size', buttonSizes, 'regular')}
       fluid={boolean('Fluid', false)}
-      modifier={select('Modifier', modifierOptions, 'hydrogen')}
-      variant={select('Variant', variantOptions, 'andromeda')}
+      modifier={select('Modifier', buttonModifiers, 'hydrogen')}
+      variant={select('Variant', buttonVariants, 'andromeda')}
       disabled={boolean('Disabled', false)}
       borderRadius={number('Border radius', 0, radiusBorderRange)}
+      fit={select('fit', buttonFitOptions, 'min-width')}
+      mobileFit={select('mobileFit', buttonMobileFitOptions, null)}
     >
       {iconPosition === 'left' && (
         <HeartIcon
@@ -195,9 +175,9 @@ WithTextAndIcon.decorators = [
 
 export const WithBadge = () => (
   <Button
-    size={select('Size', sizeOptions, 'regular')}
-    modifier={select('Modifier', modifierOptions, 'helium')}
-    variant={select('Variant', variantOptions, 'andromeda')}
+    size={select('Size', buttonSizes, 'regular')}
+    modifier={select('Modifier', buttonModifiers, 'helium')}
+    variant={select('Variant', buttonVariants, 'andromeda')}
     disabled={boolean('Disabled', false)}
   >
     <span>{text('Text', 'MyButton')}</span>